refactor: moves logic related to the pause swtich functionality to component itself for reusability
Relates #142
